+++ index.js
@@ -6,6 +6,6 @@ require('dotenv').config();
 
 // Mount routes
 // The path here must match the basePath in serverless.yml for the domain
-app.use('/past-grants-search', require('./past-grants-search'));
+app.use('/past-grants-search', require('./past-grants-search/index'));
 
 module.exports.handler = serverless(app);
